Fixed compile warning
Fixed numerous compile warnings
TEST-180 Functional tests: add --vm-name-prefix parameter
TEST-180 Functional tests: add debug printing to nat box provision script
VMS-5524 Change auto-generated certificate algorithm to SHA2 or SHA256
VMS-5553: New system is merged to cloud system incorrectly
VMS-5561 Sometimes HardwareID can't be calculated

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ unit_tests/nx_network_ut/src/buffered_stream_socket_ut.cpp
@@ -69,7 +69,7 @@ TEST_F(BufferedStreamSocketTest, catchRecvEvent)
         {
             ASSERT_EQ(code, SystemError::noError);
             ASSERT_EQ(size, kTestMessage.size() * clientCount);
-            ASSERT_EQ(buffer.size(), size);
+            ASSERT_EQ((size_t)buffer.size(), size);
             ASSERT_TRUE(buffer.startsWith(kTestMessage));
             ASSERT_TRUE(buffer.endsWith(kTestMessage));
             acceptedResults.push(code);
